backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic-settings>=2.2
python-dotenv>=1.0
httpx>=0.27

# Step 3: DB + migrations
sqlalchemy>=2.0
alembic>=1.13
asyncpg>=0.29

# Step 4: Auth
python-jose[cryptography]>=3.3
passlib>=1.7

# Step 9: Vector DB
qdrant-client>=1.12
